+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [UNRELEASED] (NAME) - YYYY-MM-DD
 ### Added
+- Added support for rfc2622 ^+, ^-, ^n and ^n-m prefix operators
 - More tests for checking withdrawn hijacks
+
 
 ### Changed
 - Updated/optimized db query for removing withdrawn peers (newer announcement)